updated to use new lib.rs layout
