+++ src-tauri/gen/schemas/capabilities.json
@@ -1,0 +1,1 @@
+{}
